Added jwt authentication
